blerf

+++ src/main/java/service/QuizREST.java
@@ -7,6 +7,7 @@ package service;
 
 import com.google.gson.Gson;
 import com.ssv.museum.core.AnswerOption;
+import com.ssv.museum.core.Museum;
 import com.ssv.museum.core.Question;
 import com.ssv.museum.core.Quiz;
 import com.ssv.museum.persistence.QuizDAO;
@@ -23,6 +24,7 @@ import javax.naming.NamingException;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
@@ -39,7 +41,7 @@ import javax.ws.rs.core.Response;
  * @author simonarneson
  */
 @Path("quiz")
-public class QuizREST {
+public class QuizREST extends AuthedREST{
 
     QuizDAO quizDAO = lookupQuizDAOBean();
     //find
@@ -70,14 +72,18 @@ public class QuizREST {
         }
     }
     //delete
+    
     @DELETE
     @Path("{id: \\d+}")
     @Produces({MediaType.APPLICATION_JSON})
     public Response delete(@PathParam("id") Long id,
+                @HeaderParam("password") String password,
                          @Context Request request) {
         //to be able to check if exists
         Quiz quiz = quizDAO.find(id);
-        if (quiz != null) {
+        Museum m = quiz.getMuseum();
+        //only the museum who added the quiz is allowed to remove it
+        if (quiz != null && m !=null && authMuseum(password,m.getUsername())) {
             quizDAO.delete(id);
             Gson gson = new Gson();
             return Response.ok(gson.toJson(quiz)).build(); // 200
@@ -87,6 +93,8 @@ public class QuizREST {
     }
 
     //create
+    /*
+    This is instead done through the museums endpoint since only museums can create quizzes
     @POST
     @Produces({MediaType.APPLICATION_JSON})
     @Consumes({MediaType.APPLICATION_JSON,MediaType.APPLICATION_XML})
@@ -103,21 +111,23 @@ public class QuizREST {
         } else {
             return Response.noContent().build();  // 204
         }
-    }
+    }*/
    //update
    @PUT
    @Path("{id: \\d+}")
    @Produces({MediaType.APPLICATION_JSON})
    @Consumes({MediaType.APPLICATION_JSON,MediaType.APPLICATION_XML})
    public Response update(@PathParam("id") Long id,
+                          @HeaderParam("password") String password,
                           JsonObject obj,@Context Request request) {
        String name = obj.getString("name");
        int points = Integer.parseInt(obj.getString("points"));
        String description = obj.getString("description");
        Quiz q = quizDAO.find(id);
-       if (q != null) {
+       Museum m = q.getMuseum();
+       if (q != null && m!=null && authMuseum(password,m.getUsername())) {
            q.setName(name);
-           q.setPoints(points);
+           q.setPoints(points); 
            q.setDescription(description);
            quizDAO.update(q);
            Gson gson = new Gson();
@@ -134,6 +144,7 @@ public class QuizREST {
     @Consumes({MediaType.APPLICATION_JSON,MediaType.APPLICATION_XML})
     public Response addQuestion(JsonObject obj,
                         @PathParam("id") Long id,
+                        @HeaderParam("password") String password,
                         @Context Request request) {
         int points = Integer.parseInt(obj.getString("points"));
         String text = obj.getString("question");
@@ -152,8 +163,9 @@ public class QuizREST {
         }
         nQuestion.setQuiz(quiz);
         quiz.addQuestion(nQuestion);
-        quizDAO.update(quiz);
-        if (quiz != null) {
+        Museum m = quiz.getMuseum();
+        if (quiz != null && m!=null && authMuseum(password,m.getUsername())) {
+            quizDAO.update(quiz);
             Gson gson = new Gson();
             return Response.ok().build();// 200
         } else {
